Map: switch to three.js (WebGL) renderer + HTML overlay split
Replaces the Canvas2D CanvasWorldMap with ThreeWorldMap. The static
layers (parchment + drop shadow + baked tile composite) render in
WebGL via three.js; the dynamic layers (party sprite, other-player
flags, count badges, hover highlight, path preview) live in a sibling
HTML overlay that carries a single CSS transform mirroring the
three.js camera. Tooltip is a separate cursor-positioned element.

Per the user's design note: the WebGL canvas just renders tiles,
shadow, and parchment; "the rest of the UI elements should sit on
top". This split makes pan/zoom essentially free (just a camera
matrix update on the GPU + one CSS transform on the overlay), and
party movement / pulse run entirely as CSS so they don't drive any
JS or WebGL work.

Render-on-demand: no permanent RAF loop. requestRender() coalesces
multiple state pushes into one frame. An anim loop runs only while
the spring-back is active. Idle map screens cost effectively zero.

Bake logic carries over unchanged (drawTile / drawZoneOverlay /
drawZoneBorders helpers, hexSpriteCache, pre-blurred shadow) — only
its output target changes from "blit to main canvas every frame" to
"upload as CanvasTexture once per bake".

- Adds three.js + @types/three to client deps
- Deletes CanvasWorldMap.ts and TileInfoModal.ts (both dead now)
- Wires MapScreen + RoomView to import TileClickInfo from
  ThreeWorldMap
- CSS for .three-world-map, .three-map-overlay, .three-map-party
  (with CSS pulse + transition for movement), .three-map-hover (SVG
  hex outline), .three-map-path-step/dest, .three-map-flag (SVG
  pole + triangle), .three-map-badge
- Docs: refresh CLAUDE.md, README.md, docs/architecture/client.md,
  docs/architecture/content.md
- Patch notes: bumped to 2026.05.13.3 (per the one-entry-per-PR rule);
  GAME_VERSION matches so the server auto-broadcasts on next boot

Phaser is still in package.json (docs already noted it was dead from
the May 2026 overhaul); cleanup is out of scope here.
