+++ client/package.json
@@ -11,9 +11,11 @@
   },
   "dependencies": {
     "@idle-party-rpg/shared": "*",
-    "phaser": "^3.80.1"
+    "phaser": "^3.80.1",
+    "three": "^0.169.0"
   },
   "devDependencies": {
+    "@types/three": "^0.169.0",
     "happy-dom": "^15.11.7",
     "typescript": "^5.3.3",
     "vite": "^5.4.2",
